Moved 'Elle.svg' to dedicated images folder.
