+++ app/src/components/submissionPage/SubmissionTextBox.tsx
@@ -11,7 +11,7 @@ import { getCurrentUserUid } from "../../firebase/users/UserService";
 import Snap from "../../types/Snap";
 import { submitSnap } from "../../firebase/snaps/SnapService";
 // @ts-ignore
-import Elle from "../loginPage/Elle.svg";
+import Elle from "../../images/Elle.svg";
 import {
     ElleImg,
     LeftSideOfScreen,
